\name{Affy1027chr2}
\alias{Affy1027chr2}
\docType{data}
\title{
TCGA sample 23-1027 Chromosome 2 from Affymetrix Platform
}
\description{
TCGA sample 23-1027 Chromosome 2 from Affymetrix Platform
}
\usage{data(Affy1027chr2)}
\format{
  A data frame with 74103 observations on the following 4 variables.
  \describe{
    \item{\code{Position}}{a numeric vector}
    \item{\code{Chr}}{a numeric vector}
    \item{\code{logR}}{a numeric vector}
    \item{\code{bfreq}}{a numeric vector}
  }
}
\examples{
data(Affy1027chr2)
}
\keyword{datasets}
