\name{AffySW1417chr17}
\alias{AffySW1417chr17}
\docType{data}
\title{
Sample SW1417 Chromosome 17 from Affymetrix Platform
}
\description{
Sample SW1417 Chromosome 17 from Affymetrix Platform.
}
\usage{data(AffySW1417chr17)}
\format{
  A data frame with 46678 observations on the following 5 variables.
  \describe{
    \item{\code{Position}}{a numeric vector}
    \item{\code{Chr}}{a numeric vector}
    \item{\code{logR}}{a numeric vector}
    \item{\code{bfreq}}{a numeric vector}
    \item{\code{probe}}{a factor with levels \code{AFFX-SNP} \code{CN} \code{SNP}}
  }
}
\examples{
data(AffySW1417chr17)
}
\keyword{datasets}
