\name{Illu0258chr2}
\alias{Illu0258chr2}
\docType{data}
\title{
TCGA sample 02-0258 Chromosome 2 from Illumina Platform
}
\description{
TCGA sample 02-0258 Chromosome 2 from Illumina Platform
}
\usage{data(Illu0258chr2)}
\format{
  A data frame with 45432 observations on the following 4 variables.
  \describe{
    \item{\code{Position}}{a numeric vector}
    \item{\code{Chr}}{a numeric vector}
    \item{\code{logR}}{a numeric vector}
    \item{\code{bfreq}}{a numeric vector}
  }
}
\examples{
data(Illu0258chr2)
}
\keyword{datasets}
