\name{Illu0332chr3}
\alias{Illu0332chr3}
\docType{data}
\title{
TCGA sample 02-0332 Chromosome 3 from Illumina Platform
}
\description{
TCGA sample 02-0332 Chromosome 3 from Illumina Platform
}
\usage{data(Illu0332chr3)}
\format{
  A data frame with 37768 observations on the following 4 variables.
  \describe{
    \item{\code{Position}}{a numeric vector}
    \item{\code{Chr}}{a numeric vector}
    \item{\code{logR}}{a numeric vector}
    \item{\code{bfreq}}{a numeric vector}
  }
}
\examples{
data(Illu0332chr3)
}
\keyword{datasets}
