\name{Illu0332chr3_mini}
\alias{Illu0332chr3_mini}
\docType{data}
\title{
TCGA sample 02-0332 Chromosome 3 from Illumina Platform
}
\description{
TCGA sample 02-0332 Chromosome 3 from Illumina Platform.  In this dataset, one SNP is used for every ten SNPs in the original dataset to make the computational time shorter.
}
\usage{data(Illu0332chr3_mini)}
\format{
  A data frame with 3777 observations on the following 4 variables.
  \describe{
    \item{\code{Position}}{a numeric vector}
    \item{\code{Chr}}{a numeric vector}
    \item{\code{logR}}{a numeric vector}
    \item{\code{bfreq}}{a numeric vector}
  }
}
\examples{
data(Illu0332chr3_mini)
}
\keyword{datasets}
